refactor: rename const file name
